Fixed issues with file owners.
Since super-user privileges are required to run the script, everything it produces is owned by root. These changes needed to be applied in order to be able to write to non-super user's '.fehbg' and not root's, therefore setting the background for the proper X sessions if '--persist' option was passed to the program.